Move value type  jni methods to their own file
